minor test refactoring to avoid conflict of test in same package
